optimizing code by sifting logic in lib
