+++ src/workstation-gui/Gemfile
@@ -83,6 +83,7 @@ gem "nokogiri"
 #   # Speed up commands on slow machines / big apps [https://github.com/rails/spring]
 #   # gem "spring"
 # end
+gem 'byebug'
 
 group :test do
   # Use system testing [https://guides.rubyonrails.org/testing.html#system-testing]
